+++ src/parsers/parseFirmwareUpdates.js
